fix: styled component Cart.jsx
