+++ src/components/Layout/InputNumber.jsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import styled from 'styled-components';
 
 function InputNumber(props) {
-  const {showNumber, setValue, maxValue, width, value} = props;
+  const {showNumber, setValue, maxValue, width, height, value} = props;
   const [qtd, setQtd] = useState(value);
 
   function increase(){
@@ -20,7 +20,7 @@ function InputNumber(props) {
   }
 
   return (
-    <ContainerInputNumber showNumber={showNumber} width={width} >
+    <ContainerInputNumber showNumber={showNumber} width={width} height={height}>
       {props.children}
       <div className="options">
         <ion-icon onClick={increase} name="add-outline"></ion-icon>
@@ -44,10 +44,11 @@ const ContainerInputNumber = styled.div`
   background-color: white;
 
   width: ${props => props.width};
+  height: ${props => props.height};
 
   border-radius: var(--radio-min);
 
-  &>.options {
+  &>div.options {
     display: flex;
     flex-direction: column;
     justify-content: space-between;
@@ -57,13 +58,14 @@ const ContainerInputNumber = styled.div`
 
     border-radius: var(--radio-min);
     background-color: var(--color-gray);
-  }
 
-  &>.options ion-icon {
-    font-size: 1.3rem;
-  }
+    ion-icon {
+      font-size: 1.3rem;
+    }
 
-  &>.options p {
-    display: var(--display-number);
+    p {
+      display: var(--display-number);
+      font-weight: bold;
+    }
   }
 `
